auto-claude: subtask-8-3 - Fix and run unit and integration tests
Fixed test failures:
- JwtTokenProvider: Added SignatureException catch for validation
- OTPServiceTest: Fixed mock verification (Integer vs Long)
- AudioProcessingServiceTest: Added user repository mocking for validation tests
- JwtTokenProviderTest: Fixed token uniqueness test with longer delay

Test Results:
- 70 unit tests pass
- 2 integration tests skipped (require Docker environment)

+++ backend/src/test/java/com/prod/service/OTPServiceTest.java
@@ -94,8 +94,8 @@ class OTPServiceTest {
         // Act
         otpService.generateAndStoreOTP(TEST_EMAIL);
 
-        // Assert
-        verify(valueOperations).set(eq("otp:" + TEST_EMAIL), anyString(), eq(OTP_EXPIRATION_MINUTES), eq(TimeUnit.MINUTES));
+        // Assert - RedisTemplate.set() takes a long for timeout, so we need to match the long value
+        verify(valueOperations).set(eq("otp:" + TEST_EMAIL), anyString(), eq((long) OTP_EXPIRATION_MINUTES), eq(TimeUnit.MINUTES));
     }
 
     @Test
